[detectors] Creating thirdeye-detectors module
