+++ thirdeye-core/pom.xml
@@ -22,6 +22,10 @@
       <artifactId>thirdeye-persistence</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.apache.pinot.thirdeye</groupId>
+      <artifactId>thirdeye-detectors</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.apache.httpcomponents</groupId>
       <artifactId>httpclient</artifactId>
     </dependency>
