+++ thirdeye-core/src/main/java/org/apache/pinot/thirdeye/detection/components/RuleBaselineProvider.java
@@ -19,19 +19,15 @@
 
 package org.apache.pinot.thirdeye.detection.components;
 
-import java.util.ArrayList;
-import java.util.List;
 import org.apache.pinot.thirdeye.dashboard.resources.v2.BaselineParsingUtils;
+import org.apache.pinot.thirdeye.detection.DetectionUtils;
 import org.apache.pinot.thirdeye.detection.spec.RuleBaselineProviderSpec;
-import org.apache.pinot.thirdeye.detection.spi.components.BaselineProvider;
-import org.apache.pinot.thirdeye.detection.spi.model.TimeSeries;
-import org.apache.pinot.thirdeye.rootcause.timeseries.Baseline;
-import org.apache.pinot.thirdeye.spi.dataframe.DataFrame;
 import org.apache.pinot.thirdeye.spi.dataframe.util.MetricSlice;
 import org.apache.pinot.thirdeye.spi.detection.InputDataFetcher;
 import org.apache.pinot.thirdeye.spi.detection.annotation.Components;
-import org.apache.pinot.thirdeye.spi.detection.spi.model.InputData;
-import org.apache.pinot.thirdeye.spi.detection.spi.model.InputDataSpec;
+import org.apache.pinot.thirdeye.spi.detection.spi.components.BaselineProvider;
+import org.apache.pinot.thirdeye.spi.detection.spi.model.TimeSeries;
+import org.apache.pinot.thirdeye.spi.rootcause.timeseries.Baseline;
 
 @Components(title = "rule baseline",
     type = "RULE_BASELINE"
@@ -45,14 +41,7 @@ public class RuleBaselineProvider implements BaselineProvider<RuleBaselineProvid
 
   @Override
   public TimeSeries computePredictedTimeSeries(MetricSlice slice) {
-    return TimeSeries.fromDataFrame(buildBaselines(slice, this.baseline, this.dataFetcher));
-  }
-
-  static DataFrame buildBaselines(MetricSlice slice, Baseline baseline,
-      InputDataFetcher dataFetcher) {
-    List<MetricSlice> slices = new ArrayList<>(baseline.scatter(slice));
-    InputData data = dataFetcher.fetchData(new InputDataSpec().withTimeseriesSlices(slices));
-    return baseline.gather(slice, data.getTimeseries());
+    return TimeSeries.fromDataFrame(DetectionUtils.buildBaselines(slice, this.baseline, this.dataFetcher));
   }
 
   @Override
