+++ thirdeye-core/src/main/java/org/apache/pinot/thirdeye/detection/spec/MeanVarianceRuleDetectorSpec.java
@@ -19,8 +19,8 @@
 
 package org.apache.pinot.thirdeye.detection.spec;
 
-import org.apache.pinot.thirdeye.detection.Pattern;
 import org.apache.pinot.thirdeye.spi.dataframe.util.MetricSlice;
+import org.apache.pinot.thirdeye.spi.detection.Pattern;
 import org.apache.pinot.thirdeye.spi.detection.spec.AbstractSpec;
 
 public class MeanVarianceRuleDetectorSpec extends AbstractSpec {
