+++ thirdeye-deprecated-tests/src/test/java/org/apache/pinot/thirdeye/detection/components/MockTunableDetector.java
@@ -24,12 +24,12 @@ package org.apache.pinot.thirdeye.detection.components;
 
 import java.util.Collections;
 import java.util.Map;
-import org.apache.pinot.thirdeye.spi.detection.InputDataFetcher;
 import org.apache.pinot.thirdeye.detection.spec.MockTunableSpec;
-import org.apache.pinot.thirdeye.detection.spi.components.AnomalyDetector;
 import org.apache.pinot.thirdeye.detection.spi.components.Tunable;
+import org.apache.pinot.thirdeye.spi.detection.InputDataFetcher;
+import org.apache.pinot.thirdeye.spi.detection.spi.components.AnomalyDetector;
 import org.apache.pinot.thirdeye.spi.detection.spi.exception.DetectorException;
-import org.apache.pinot.thirdeye.detection.spi.model.DetectionResult;
+import org.apache.pinot.thirdeye.spi.detection.spi.model.DetectionResult;
 import org.joda.time.Interval;
 
 public class MockTunableDetector implements AnomalyDetector<MockTunableSpec>,
