+++ thirdeye-deprecated-tests/src/test/java/org/apache/pinot/thirdeye/detection/spec/MockBaselineProviderSpec.java
@@ -19,7 +19,7 @@ package org.apache.pinot.thirdeye.detection.spec;
 import java.util.Map;
 import org.apache.pinot.thirdeye.spi.dataframe.util.MetricSlice;
 import org.apache.pinot.thirdeye.spi.detection.spec.AbstractSpec;
-import org.apache.pinot.thirdeye.detection.spi.model.TimeSeries;
+import org.apache.pinot.thirdeye.spi.detection.spi.model.TimeSeries;
 
 public class MockBaselineProviderSpec extends AbstractSpec {
