+++ thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/spi/detection/spi/components/BaselineProvider.java
@@ -17,13 +17,12 @@
  * under the License.
  */
 
-package org.apache.pinot.thirdeye.detection.spi.components;
+package org.apache.pinot.thirdeye.spi.detection.spi.components;
 
-import org.apache.pinot.thirdeye.detection.spi.model.TimeSeries;
 import org.apache.pinot.thirdeye.spi.dataframe.Series;
 import org.apache.pinot.thirdeye.spi.dataframe.util.MetricSlice;
 import org.apache.pinot.thirdeye.spi.detection.spec.AbstractSpec;
-import org.apache.pinot.thirdeye.spi.detection.spi.components.BaseComponent;
+import org.apache.pinot.thirdeye.spi.detection.spi.model.TimeSeries;
 
 /**
  * The baseline provider to calculate predicted baseline.
