+++ thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/spi/detection/spi/model/DetectionResult.java
@@ -19,7 +19,7 @@
  *
  *
  */
-package org.apache.pinot.thirdeye.detection.spi.model;
+package org.apache.pinot.thirdeye.spi.detection.spi.model;
 
 import java.util.Collections;
 import java.util.List;
